refactor: checking for the start date before return matches and check if ther is stadium on db
