--- client/src/pages/.gitkeep
@@ -1,1 +1,0 @@
-gitkeep
