--- server/services/matches.ts
@@ -1,9 +1,0 @@
-import { Request } from 'express';
-import { Match } from '../models';
-
-export const getStadiumMatchesService = async (
-  req: Request,
-): Promise<Match[]> => {
-  const { stadiumId } = req.params;
-  return await Match.findAll({ where: { UserId: stadiumId } });
-};
